pyproject.toml
[project]
name = "messagelane"
dynamic = ["version"]
description = "SQL Message Queue"
readme = "README.rst"
license = { file = "LICENSE" }
requires-python= ">= 3.9"
authors = [
    { name= 'Todd Valentic', email='todd.valentic@gmail.com' }
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

dependencies = [
    "click",
    "python-dotenv",
    "psycopg2-binary",
    "sqlalchemy>=2",
    "texttable",
    "prefixed",
    "importlib_resources>1.3; python_version < '3.10'"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
devel = [
    'ruff',
    'pytest',
    'pyfakefs'
    ]

[project.urls]
homepage="https://github.com/valentic/messagelane"
repository="https://github.com/valentic/messagelane"

[project.scripts]
    msglane = "messagelane.commands.msglane:main"

#[tool.setuptools.packages.find]
#where = ["src"]

[tool.hatch.version]
path = "src/messagelane/metadata.py"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E", 
    # pyflakes
    "F", 
    # flake8-bugbear
    "B", 
    # isort
    # I"
    # flake8-simplify
    "SIM", 
    # docstyle
    "D"
]
